+++ src/errors.rs
@@ -1,13 +1,29 @@
+use regex::Error;
 use serde_json;
 use std::io;
 use std::num::ParseIntError;
-use regex::Error;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
 pub enum CLIError {
-    #[error("failed to run linter")]
-    LintError(#[from] LintError)
+    #[error("failed to initialize the changelog settings: {0}")]
+    InitError(#[from] InitError),
+    #[error("failed to run linter: {0}")]
+    LintError(#[from] LintError),
+    #[error("failed to read configuration: {0}")]
+    Config(#[from] ConfigError),
+    #[error("failed to adjust configuration: {0}")]
+    ConfigAdjustment(#[from] ConfigAdjustError),
+    #[error("failed to read/write: {0}")]
+    IOError(#[from] io::Error),
+}
+
+#[derive(Error, Debug)]
+pub enum InitError {
+    #[error("failed to write: {0}")]
+    FailedToWrite(#[from] io::Error),
+    #[error("config already created")]
+    ConfigAlreadyFound,
 }
 
 #[derive(Error, Debug)]
@@ -58,8 +74,24 @@ pub enum MatchError {
 
 #[derive(Error, Debug)]
 pub enum ConfigError {
+    #[error("failed to read/write configuration: {0}")]
+    FailedToReadWrite(#[from] io::Error),
     #[error("failed to parse configuration")]
     FailedToParse(#[from] serde_json::Error),
+}
+
+#[derive(Error, Debug, PartialEq)]
+pub enum ConfigAdjustError {
+    #[error("category already found")]
+    CategoryAlreadyFound,
+    #[error("key is already present in hash map")]
+    KeyAlreadyFound,
+    #[error("Invalid URL")]
+    InvalidURL(#[from] url::ParseError),
+    #[error("expected value not found")]
+    NotFound,
+    #[error("target repository should be a GitHub link")]
+    NoGitHubRepository,
 }
 
 #[derive(Error, Debug, PartialEq)]