+++ src/lib.rs
@@ -1,9 +1,11 @@
 mod change_type;
 mod changelog;
 pub mod cli;
+pub mod cli_config;
 pub mod config;
 mod entry;
 pub mod errors;
+pub mod init;
 pub mod lint;
 mod release;
 mod version;