+++ src/lint.rs
@@ -19,8 +19,7 @@ pub fn run(fix: bool) -> Result<(), LintError> {
         }
     };
 
-    // TODO: check for configuration file in user directory
-    let config = config::load(
+    let config = config::unpack_config(
         fs::read_to_string(Path::new(".clconfig.json"))?.as_str()
     )?;